Move map_item helper scripts into their own folder
